refactor: Re-arrange code for readability
