+++ Sources/Confidence/Client/ConfidenceClientResolveRemote.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public class RemoteConfidenceResolveClient: ConfidenceResolveClient {
+public class ConfidenceClientResolveRemote: ConfidenceResolveClient {
     private let targetingKey = "targeting_key"
     private var options: ConfidenceClientOptions
     private let metadata: ConfidenceMetadata
