--- Sources/Confidence/Sdk.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public struct Sdk: Codable {
-    var id: String
-    var version: String
-}
