+++ Sources/Confidence/Utils/BaseUrlMapper.swift
