+++ Sources/Confidence/Utils/ConfidenceValueHash.swift
