+++ Sources/Confidence/Utils/HttpStatusCode+Error.swift
