+++ Sources/Confidence/Utils/VisitorUtil.swift
