+++ Tests/ConfidenceTests/ConfidenceTest.swift
@@ -591,4 +591,17 @@ final class DispatchQueueFake: DispatchQueueType {
         work()
     }
 }
+
+
+/// Used for testing
+public protocol DispatchQueueType {
+    func async(execute work: @escaping @convention(block) () -> Void)
+}
+
+extension DispatchQueue: DispatchQueueType {
+    public func async(execute work: @escaping @convention(block) () -> Void) {
+        async(group: nil, qos: .unspecified, flags: [], execute: work)
+    }
+}
+
 // swiftlint:enable type_body_length
